+++ app/src/main/java/com/brandonjamesyoung/levelup/fragments/NewQuest.kt
@@ -1,6 +1,8 @@
 package com.brandonjamesyoung.levelup.fragments
 
 import android.os.Bundle
+import android.text.Editable
+import android.text.TextWatcher
 import android.util.Log
 import android.view.View
 import android.widget.EditText
@@ -10,7 +12,6 @@ import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.constraintlayout.widget.ConstraintSet
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.fragment.navArgs
@@ -18,6 +19,7 @@ import com.brandonjamesyoung.levelup.R
 import com.brandonjamesyoung.levelup.data.Quest
 import com.brandonjamesyoung.levelup.shared.ByteArrayHelper.Companion.convertByteArrayToDrawable
 import com.brandonjamesyoung.levelup.shared.Difficulty
+import com.brandonjamesyoung.levelup.shared.IconHelper.Companion.getDefaultIcon
 import com.brandonjamesyoung.levelup.shared.Mode
 import com.brandonjamesyoung.levelup.viewmodels.NewQuestViewModel
 import com.brandonjamesyoung.levelup.validation.Validation.Companion.validateName
@@ -28,8 +30,6 @@ import kotlinx.coroutines.launch
 @AndroidEntryPoint
 class NewQuest : Fragment(R.layout.new_quest) {
     private val viewModel: NewQuestViewModel by activityViewModels()
-    private var selectedDifficulty: Difficulty? = null
-    private var mode: MutableLiveData<Mode> = MutableLiveData<Mode>()
     private val args: NewQuestArgs by navArgs()
 
     private val buttonIdToDifficultyMap = mapOf(
@@ -41,6 +41,28 @@ class NewQuest : Fragment(R.layout.new_quest) {
 
     private val difficultyToButtonIdMap = buttonIdToDifficultyMap.entries
         .associateBy({ it.value }) { it.key }
+
+    private fun setupNameInput() {
+        val view = requireView()
+        val nameInput = view.findViewById<EditText>(R.id.NameInput)
+
+        if (viewModel.name != null) {
+            nameInput.setText(viewModel.name)
+        }
+
+        nameInput.addTextChangedListener(object : TextWatcher {
+            override fun afterTextChanged(s: Editable) {
+            }
+
+            override fun beforeTextChanged(s: CharSequence, start: Int, count: Int, after: Int) {
+            }
+
+            override fun onTextChanged(s: CharSequence, start: Int, before: Int, count: Int) {
+                val name = s.toString()
+                viewModel.name = name.ifEmpty { null }
+            }
+        })
+    }
 
     private fun moveDifficultySelectBox(difficulty: Difficulty) {
         val view = requireView()
@@ -58,9 +80,9 @@ class NewQuest : Fragment(R.layout.new_quest) {
     }
 
     private fun setSelectedDifficulty(difficulty: Difficulty) {
-        selectedDifficulty = difficulty
+        viewModel.selectedDifficulty = difficulty
         moveDifficultySelectBox(difficulty)
-        Log.i(TAG, "Set quest difficulty to $selectedDifficulty")
+        Log.i(TAG, "Set quest difficulty to $difficulty")
     }
 
     private fun setDifficultyButtonListeners() {
@@ -75,6 +97,17 @@ class NewQuest : Fragment(R.layout.new_quest) {
         }
     }
 
+    private fun setupDifficultyInput() {
+        setDifficultyButtonListeners()
+        setSelectedDifficulty(viewModel.selectedDifficulty)
+    }
+
+    private fun setupInputFields() {
+        setupNameInput()
+        setupDifficultyInput()
+        setupIconSelectButton()
+    }
+
     private fun validateInput() : Boolean {
         val view = requireView()
         val nameInput = view.findViewById<EditText>(R.id.NameInput)
@@ -83,37 +116,10 @@ class NewQuest : Fragment(R.layout.new_quest) {
             return false
         }
 
-        // TODO Check if icon is a valid icon name
+        // TODO Check if icon is a valid id
         // val iconView = view.findViewById<FloatingActionButton>(R.id.IconButton)
 
         return true
-    }
-
-    private fun saveQuest() {
-        val view = requireView()
-        val nameTextView = view.findViewById<EditText>(R.id.NameInput)
-        var name: String? = nameTextView.text.trim().toString()
-
-        if (name == "") {
-            name = null
-        }
-
-        // TODO get icon file name here and store in saved quest
-//        val iconButton = view.findViewById<FloatingActionButton>(R.id.IconButton)
-//        val iconDrawable = iconButton.drawable
-//        val iconFileName = resources.getResourceEntryName(R.drawable.question_mark_icon)
-
-        val quest = Quest(
-            name = name,
-            difficulty = selectedDifficulty!!,
-        )
-
-        if (mode.value == Mode.DEFAULT) {
-            viewModel.insert(quest)
-        } else if (mode.value == Mode.EDIT) {
-            quest.id = args.questId
-            viewModel.update(quest)
-        }
     }
 
     private fun setupConfirmButton() {
@@ -122,13 +128,14 @@ class NewQuest : Fragment(R.layout.new_quest) {
 
         saveButton.setOnClickListener {
             if (validateInput()){
-                saveQuest()
+                viewModel.saveQuest()
                 navigateToQuestList()
             }
         }
     }
 
     private fun navigateToQuestList() {
+        viewModel.clearInput()
         NavHostFragment.findNavController(this).navigate(R.id.action_newQuest_to_questList)
         Log.i(TAG, "Going from New Quest to Quest List")
     }
@@ -147,15 +154,37 @@ class NewQuest : Fragment(R.layout.new_quest) {
         Log.i(TAG, "Going from New Quest to Icon Select")
     }
 
+    private fun changeIcon(iconId : Int?) {
+        val view = requireView()
+        val button = view.findViewById<FloatingActionButton>(R.id.IconButton)
+
+        // TODO add check for invalid id, if invalid then show default icon
+        if (iconId == null) {
+            val drawable = getDefaultIcon(view.context)
+            button.setImageDrawable(drawable)
+            return
+        }
+
+        val iconLiveData = viewModel.getIcon(iconId)
+
+        iconLiveData.observe(viewLifecycleOwner) { icon ->
+            val drawable = convertByteArrayToDrawable(icon.image, resources)
+            button.setImageDrawable(drawable)
+            iconLiveData.removeObservers(viewLifecycleOwner)
+        }
+    }
+
     private fun setupIconSelectButton() {
         val view = requireView()
         val button = view.findViewById<FloatingActionButton>(R.id.IconButton)
 
+        if (viewModel.iconId != null) {
+            changeIcon(viewModel.iconId as Int)
+        }
+
         if (args.iconId != INVALID_ICON_ID) {
-            viewModel.getIcon(args.iconId).observe(viewLifecycleOwner) { icon ->
-                val drawable = convertByteArrayToDrawable(icon.image, resources)
-                button.setImageDrawable(drawable)
-            }
+            viewModel.iconId = args.iconId
+            changeIcon(args.iconId)
         }
 
         button.setOnClickListener{
@@ -164,14 +193,8 @@ class NewQuest : Fragment(R.layout.new_quest) {
     }
 
     private fun setupButtons() {
-        setDifficultyButtonListeners()
         setupConfirmButton()
         setupCancelButton()
-        setupIconSelectButton()
-    }
-
-    private fun selectDefaultDifficulty() {
-        setSelectedDifficulty(DEFAULT_DIFFICULTY)
     }
 
     private fun loadQuest(quest: Quest) {
@@ -179,6 +202,7 @@ class NewQuest : Fragment(R.layout.new_quest) {
         val nameInput = view.findViewById<EditText>(R.id.NameInput)
         nameInput.setText(quest.name)
         setSelectedDifficulty(quest.difficulty)
+        changeIcon(quest.iconId)
     }
 
     private fun activateEditMode() {
@@ -186,19 +210,20 @@ class NewQuest : Fragment(R.layout.new_quest) {
         val pageLabel = view.findViewById<TextView>(R.id.NewQuestLabel)
         pageLabel.text = resources.getString(R.string.edit_quest_label)
 
-        viewModel.getQuest(args.questId).observe(viewLifecycleOwner) { quest ->
+        viewModel.getQuest(viewModel.editQuestId as Int).observe(viewLifecycleOwner) { quest ->
             loadQuest(quest)
         }
     }
 
     private fun setupMode() {
-        mode.value = Mode.DEFAULT
+        viewModel.mode.value = Mode.DEFAULT
 
         if (args.questId != INVALID_QUEST_ID) {
-            mode.value = Mode.EDIT
+            viewModel.mode.value = Mode.EDIT
+            viewModel.editQuestId = args.questId
         }
 
-        mode.observe(viewLifecycleOwner) { mode ->
+        viewModel.mode.observe(viewLifecycleOwner) { mode ->
             when (mode) {
                 Mode.DEFAULT -> Unit
                 Mode.EDIT -> activateEditMode()
@@ -212,8 +237,8 @@ class NewQuest : Fragment(R.layout.new_quest) {
 
         lifecycleScope.launch {
             Log.i(TAG, "On New Quest page")
+            setupInputFields()
             setupButtons()
-            selectDefaultDifficulty()
             setupMode()
         }
     }
@@ -222,6 +247,5 @@ class NewQuest : Fragment(R.layout.new_quest) {
         private const val TAG = "NewQuest"
         private const val INVALID_QUEST_ID = 0
         private const val INVALID_ICON_ID = 0
-        private val DEFAULT_DIFFICULTY = Difficulty.EASY
     }
 }